agent-server/fastapi/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.12.1
python-dotenv>=1.0.0
pydantic>=2.8.0
pydantic-settings>=2.8.0
langchain==0.1.0
langchain-openai==0.0.2
langgraph==0.0.20
openai>=1.6.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
httpx==0.25.2